Add `AuthorizationApiPool` trait
